626 NumField: moves back to main folder since there is no localized styles for this component
