--- packages/orchestrator-ui-components/src/components/WfoForms/formFields/NumField/index.ts
@@ -1,1 +1,0 @@
-export * from './NumField';
